+++ rspress.config.ts
@@ -31,6 +31,8 @@ export default defineConfig({
   markdown: {
     showLineNumbers: true,
   },
-  globalUIComponents: [path.join(__dirname, 'src', 'components', 'Terminal.tsx')],
-  globalStyles: path.join(__dirname, 'src', 'styles', 'terminal.css'),
+  globalUIComponents: [
+    path.join(__dirname, 'src', 'components', 'Terminal.tsx'),
+  ],
+  globalStyles: path.join(__dirname, 'src', 'styles', 'global.css'),
 });