+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### Feat
 
-- **Terminal*.tsx**: Add Terminal components for better UX
+- **Terminal\*.tsx**: Add Terminal components for better UX
 - **Introduction-and-Gossip**: Add structure and fill Introduction and Gossip section
 
 ### Refactor
 
-- **database_*.md**: Rename files
+- **database\_\*.md**: Rename files